+++ packages/auth/src/core/auth/emulator.ts
@@ -18,7 +18,12 @@ import { Auth } from '../../model/public_types';
 import { AuthErrorCode } from '../errors';
 import { _assert } from '../util/assert';
 import { _castAuth } from './auth_impl';
-import { deepEqual } from '@firebase/util';
+import {
+  deepEqual,
+  isCloudWorkstation,
+  pingServer,
+  updateEmulatorBanner
+} from '@firebase/util';
 
 /**
  * Changes the {@link Auth} instance to communicate with the Firebase Auth Emulator, instead of production
@@ -97,7 +102,11 @@ export function connectAuthEmulator(
   authInternal.emulatorConfig = emulatorConfig;
   authInternal.settings.appVerificationDisabledForTesting = true;
 
-  if (!disableWarnings) {
+  if (isCloudWorkstation(host)) {
+    updateEmulatorBanner('Auth', true);
+    // Workaround to get cookies in Firebase Studio
+    void pingServer(`${protocol}//${host}${portStr}`);
+  } else if (!disableWarnings) {
     emitEmulatorWarning();
   }
 }