+++ packages/firestore/test/integration/api/query.test.ts
@@ -109,7 +109,6 @@ apiDescribe('Queries', persistence => {
       if (path !== null) {
         await withTestDb(persistence, async db => {
           const docRef = doc(db, path!);
-          console.error('DEDB loading bundle with json: ', json);
           await loadBundle(db, json);
           const docSnap = await getDocFromCache(docRef);
           expect(docSnap.exists);
@@ -1402,6 +1401,7 @@ apiDescribe('Queries', persistence => {
       return withTestCollection(persistence, testDocs, async coll => {
         // a == 1
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, where('a', '==', 1)),
           'doc1',
           'doc4',
@@ -1410,18 +1410,21 @@ apiDescribe('Queries', persistence => {
 
         // Implicit AND: a == 1 && b == 3
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, where('a', '==', 1), where('b', '==', 3)),
           'doc4'
         );
 
         // explicit AND: a == 1 && b == 3
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, and(where('a', '==', 1), where('b', '==', 3))),
           'doc4'
         );
 
         // a == 1, limit 2
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, where('a', '==', 1), limit(2)),
           'doc1',
           'doc4'
@@ -1429,6 +1432,7 @@ apiDescribe('Queries', persistence => {
 
         // explicit OR: a == 1 || b == 1 with limit 2
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, or(where('a', '==', 1), where('b', '==', 1)), limit(2)),
           'doc1',
           'doc2'
@@ -1436,6 +1440,7 @@ apiDescribe('Queries', persistence => {
 
         // only limit 2
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, limit(2)),
           'doc1',
           'doc2'
@@ -1443,6 +1448,7 @@ apiDescribe('Queries', persistence => {
 
         // limit 2 and order by b desc
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, limit(2), orderBy('b', 'desc')),
           'doc4',
           'doc3'
@@ -1462,6 +1468,7 @@ apiDescribe('Queries', persistence => {
       return withTestCollection(persistence, testDocs, async coll => {
         // Two equalities: a==1 || b==1.
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, or(where('a', '==', 1), where('b', '==', 1))),
           'doc1',
           'doc2',
@@ -1471,6 +1478,7 @@ apiDescribe('Queries', persistence => {
 
         // (a==1 && b==0) || (a==3 && b==2)
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             or(
@@ -1484,6 +1492,7 @@ apiDescribe('Queries', persistence => {
 
         // a==1 && (b==0 || b==3).
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             and(
@@ -1497,6 +1506,7 @@ apiDescribe('Queries', persistence => {
 
         // (a==2 || b==2) && (a==3 || b==3)
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             and(
@@ -1509,6 +1519,7 @@ apiDescribe('Queries', persistence => {
 
         // Test with limits without orderBy (the __name__ ordering is the tie breaker).
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, or(where('a', '==', 2), where('b', '==', 1)), limit(1)),
           'doc2'
         );
@@ -1528,6 +1539,7 @@ apiDescribe('Queries', persistence => {
       return withTestCollection(persistence, testDocs, async coll => {
         // a==2 || b in [2,3]
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, or(where('a', '==', 2), where('b', 'in', [2, 3]))),
           'doc3',
           'doc4',
@@ -1549,6 +1561,7 @@ apiDescribe('Queries', persistence => {
       return withTestCollection(persistence, testDocs, async coll => {
         // a==2 || b array-contains 7
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, or(where('a', '==', 2), where('b', 'array-contains', 7))),
           'doc3',
           'doc4',
@@ -1557,6 +1570,7 @@ apiDescribe('Queries', persistence => {
 
         // a==2 || b array-contains-any [0, 3]
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             or(where('a', '==', 2), where('b', 'array-contains-any', [0, 3]))
@@ -1580,6 +1594,7 @@ apiDescribe('Queries', persistence => {
 
       return withTestCollection(persistence, testDocs, async coll => {
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             or(
@@ -1594,6 +1609,7 @@ apiDescribe('Queries', persistence => {
         );
 
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             and(
@@ -1605,6 +1621,7 @@ apiDescribe('Queries', persistence => {
         );
 
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             or(
@@ -1618,6 +1635,7 @@ apiDescribe('Queries', persistence => {
         );
 
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             and(
@@ -1643,6 +1661,7 @@ apiDescribe('Queries', persistence => {
 
       return withTestCollection(persistence, testDocs, async coll => {
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             or(where('a', 'in', [2, 3]), where('b', 'array-contains', 3))
@@ -1653,6 +1672,7 @@ apiDescribe('Queries', persistence => {
         );
 
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             and(where('a', 'in', [2, 3]), where('b', 'array-contains', 7))
@@ -1661,6 +1681,7 @@ apiDescribe('Queries', persistence => {
         );
 
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             or(
@@ -1674,6 +1695,7 @@ apiDescribe('Queries', persistence => {
         );
 
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             and(
@@ -1698,6 +1720,7 @@ apiDescribe('Queries', persistence => {
 
       return withTestCollection(persistence, testDocs, async coll => {
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, where('a', '==', 1), orderBy('a')),
           'doc1',
           'doc4',
@@ -1705,6 +1728,7 @@ apiDescribe('Queries', persistence => {
         );
 
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, where('a', 'in', [2, 3]), orderBy('a')),
           'doc6',
           'doc3'
@@ -1725,6 +1749,7 @@ apiDescribe('Queries', persistence => {
       return withTestCollection(persistence, testDocs, async coll => {
         // Two IN operations on different fields with disjunction.
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, or(where('a', 'in', [2, 3]), where('b', 'in', [0, 2]))),
           'doc1',
           'doc3',
@@ -1733,6 +1758,7 @@ apiDescribe('Queries', persistence => {
 
         // Two IN operations on different fields with conjunction.
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, and(where('a', 'in', [2, 3]), where('b', 'in', [0, 2]))),
           'doc3'
         );
@@ -1740,6 +1766,7 @@ apiDescribe('Queries', persistence => {
         // Two IN operations on the same field.
         // a IN [1,2,3] && a IN [0,1,4] should result in "a==1".
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             and(where('a', 'in', [1, 2, 3]), where('a', 'in', [0, 1, 4]))
@@ -1752,6 +1779,7 @@ apiDescribe('Queries', persistence => {
         // a IN [2,3] && a IN [0,1,4] is never true and so the result should be an
         // empty set.
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             and(where('a', 'in', [2, 3]), where('a', 'in', [0, 1, 4]))
@@ -1760,6 +1788,7 @@ apiDescribe('Queries', persistence => {
 
         // a IN [0,3] || a IN [0,2] should union them (similar to: a IN [0,2,3]).
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(coll, or(where('a', 'in', [0, 3]), where('a', 'in', [0, 2]))),
           'doc3',
           'doc6'
@@ -1767,6 +1796,7 @@ apiDescribe('Queries', persistence => {
 
         // Nested composite filter on the same field.
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             and(
@@ -1782,6 +1812,7 @@ apiDescribe('Queries', persistence => {
 
         // Nested composite filter on the different fields.
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           query(
             coll,
             and(
@@ -1817,6 +1848,7 @@ apiDescribe('Queries', persistence => {
 
         let testQuery = query(coll, where('zip', '!=', 98101));
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           testQuery,
           'a',
           'b',
@@ -1829,6 +1861,7 @@ apiDescribe('Queries', persistence => {
 
         testQuery = query(coll, where('zip', '!=', Number.NaN));
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           testQuery,
           'b',
           'c',
@@ -1841,6 +1874,7 @@ apiDescribe('Queries', persistence => {
 
         testQuery = query(coll, where('zip', '!=', null));
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           testQuery,
           'a',
           'b',
@@ -1877,6 +1911,7 @@ apiDescribe('Queries', persistence => {
           where('zip', 'not-in', [98101, 98103, [98101, 98102]])
         );
         await checkOnlineAndOfflineResultsMatch(
+          coll,
           testQuery,
           'a',
           'b',
@@ -1887,7 +1922,7 @@ apiDescribe('Queries', persistence => {
         );
 
         testQuery = query(coll, where('zip', 'not-in', [null]));
-        await checkOnlineAndOfflineResultsMatch(testQuery);
+        await checkOnlineAndOfflineResultsMatch(coll, testQuery);
       });
     });
   });