+++ packages/functions/src/api.ts
@@ -29,7 +29,8 @@ import {
 } from './service';
 import {
   getModularInstance,
-  getDefaultEmulatorHostnameAndPort
+  getDefaultEmulatorHostnameAndPort,
+  updateEmulatorBanner
 } from '@firebase/util';
 
 export { FunctionsError } from './error';
@@ -47,6 +48,7 @@ export function getFunctions(
   app: FirebaseApp = getApp(),
   regionOrCustomDomain: string = DEFAULT_REGION
 ): Functions {
+  updateEmulatorBanner('Functions', false);
   // Dependencies
   const functionsProvider: Provider<'functions'> = _getProvider(
     getModularInstance(app),