+++ packages/storage/src/service.ts
@@ -42,7 +42,13 @@ import {
 } from './implementation/error';
 import { validateNumber } from './implementation/type';
 import { FirebaseStorage } from './public-types';
-import { createMockUserToken, EmulatorMockTokenOptions } from '@firebase/util';
+import {
+  createMockUserToken,
+  EmulatorMockTokenOptions,
+  isCloudWorkstation,
+  pingServer,
+  updateEmulatorBanner
+} from '@firebase/util';
 import { Connection, ConnectionType } from './implementation/connection';
 
 export function isUrl(path?: string): boolean {
@@ -141,7 +147,14 @@ export function connectStorageEmulator(
   } = {}
 ): void {
   storage.host = `${host}:${port}`;
-  storage._protocol = 'http';
+  const useSsl = isCloudWorkstation(host);
+  // Workaround to get cookies in Firebase Studio
+  if (useSsl) {
+    void pingServer(`https://${storage.host}`);
+    updateEmulatorBanner('Storage', true);
+  }
+  storage._isUsingEmulator = true;
+  storage._protocol = useSsl ? 'https' : 'http';
   const { mockUserToken } = options;
   if (mockUserToken) {
     storage._overrideAuthToken =
@@ -187,7 +200,8 @@ export class FirebaseStorageImpl implements FirebaseStorage {
      * @internal
      */
     readonly _url?: string,
-    readonly _firebaseVersion?: string
+    readonly _firebaseVersion?: string,
+    public _isUsingEmulator = false
   ) {
     this._maxOperationRetryTime = DEFAULT_MAX_OPERATION_RETRY_TIME;
     this._maxUploadRetryTime = DEFAULT_MAX_UPLOAD_RETRY_TIME;
@@ -320,7 +334,8 @@ export class FirebaseStorageImpl implements FirebaseStorage {
         appCheckToken,
         requestFactory,
         this._firebaseVersion,
-        retry
+        retry,
+        this._isUsingEmulator
       );
       this._requests.add(request);
       // Request removes itself from set when complete.